Correctly 'continue' requests
'preq' modifies the request object passed to it. This results in
'this.options' also getting modified as 'extend',

    "Extends one object (the target object. Here, 'this .options')
     with one or more others, returning the modified object."

Creating the request object from the modified 'this.options' property
in the successive requests results in the query not continuing as it should.

So create a request object with a copy of 'this.options' and thus continuing
the query as it actually should.